sw_apps: versal_plm: fixing misra c violation for rule 10.4 in xplm_loader.c
Fixed rule 10.4 by typecasting where required.


Acked-for-series: Vikram Sreenivasa Batchali <b.vikram@xilinx.com>

+++ lib/sw_apps/versal_plm/src/xplm_loader.c
@@ -101,7 +101,7 @@ int XPlm_LoadBootPdi(void *Arg)
 	 * no PDI gets loaded
 	 */
 	if ((BootMode == XLOADER_PDI_SRC_JTAG) &&
-	    (XLoader_IsJtagSbiMode() == FALSE)) {
+	    (XLoader_IsJtagSbiMode() == (u8)FALSE)) {
 		Status = XST_SUCCESS;
 		goto END;
 	}
